Make data type char* to const char * in yt_particle, libyt only borrows their names.
libyt only borrows defined names (everything about char*) from simulation, and does not make a copy.
I'm not sure do we really need to make a copy, since everything will be inside python, and if we really don't want to care about the lifetime of these char *, we can always look up inside python.
But I think it is unnecessary to make a copy or look it up inside python, since these are variables already defined in simulation. If not, then we can always assign it to a const char *.