Remove flag file if it's create by libyt, and remove reload instruction file.
After reading <reload>/<reload>_EXIT file to instruct it to reload or to exit, libyt removes the file.
Also, if flag file to enter reload script mode is generated by libyt because there is an error in inline function, then the flag file will be remove by libyt when exit the API.
Flag file indicates a state.