This avoids parsing the wrong file to display the error.
Py_CompileString(..., ..., Py_file_input) parses the file to display the traceback error. I don't know why if it cannot get the file, then it uses the inline script import. Thus, it is displaying the wrong traceback, because it is parsing the wrong file.
To avoid that, I set the file name to <In [1]>, so when executing codes in Jupyter, it is displaying the error correctly.