[Unit Test -- DS Amr] Testing binding heirarchy, numpy part is not working.
While example can run successfully, the unit test failed at numpy part.
I think this is because the initialization of numpy cannot be shared between the one in libyt library and the one initialized in googletest. Because I can successfully called numpy c api in unit test, but failed in libyt.