[Unit Test] Fix numpy issue when unit testing.
This is because NumPy API initialization only imports the api within a translation unit.
Since we compile it to a libyt library and make unit test link to it, it is in separate translation unit, we need to have a public API to initialize NumPy API within libyt library itself.
(https://stackoverflow.com/questions/32899621/numpy-capi-error-with-import-array-when-compiling-multiple-modules/35362918)